added posts to group app
