+++ peer_mentoring/groups/models.py
@@ -1,8 +1,6 @@
-from django.db import models
-
 from django.contrib.auth import get_user_model
-
-
+from django.contrib.auth.models import User
+from django.db import models
 
 # Create your models here.
 
@@ -12,7 +10,7 @@ user = get_user_model()
 class Avatar(models.Model):
     id = models.BigAutoField(primary_key=True)
     name = models.CharField(max_length=30, blank=False)
-    image = models.ImageField(upload_to='avatar')
+    image = models.ImageField(upload_to="avatar")
 
     def __str__(self):
         return self.name
@@ -35,3 +33,27 @@ class Group(models.Model):
 
     def __str__(self):
         return self.title
+
+
+class Post(models.Model):
+    id = models.BigAutoField(primary_key=True)
+    author = models.ForeignKey(User, on_delete=models.CASCADE)
+    title = models.CharField(max_length=150)
+    content = models.TextField(blank=False, null=True)
+    added = models.DateTimeField(auto_now_add=True, null=False, blank=False)
+    group = models.ForeignKey(Group, on_delete=models.CASCADE, blank=True, null=True)
+
+    @property
+    def is_group_post(self):
+        return self.group is not None
+
+    @is_group_post.setter
+    def group_post_id(self, new_value: int):
+        self.group = new_value
+
+
+class Comment(models.Model):
+    author = models.ForeignKey(User, on_delete=models.CASCADE)
+    content = models.TextField(max_length=200, blank=False, null=True)
+    added = models.DateTimeField(auto_now_add=True, null=False, blank=False)
+    post = models.ForeignKey(Post, on_delete=models.CASCADE)
